Improved plugin system
You can now set a priority on a function to control
when it executes for a particular action (hook).
Default priority is 10, but can be changed to anything.

For example, if you have 2 functions hooked to 'Phantom',
you can change one function to priority 12 and leave the other
at 10, so the one at 10 will execute first, no matter where
it is defined.

old-commit-id: ce065d9fda4d26aed666808925247807bd174234